requirements.txt
Flask

twilio
sendgrid

Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
python-dotenv

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai

marshmallow
flask-marshmallow
marshmallow-sqlalchemy

flask-migrate
psycopg2-binary
